Add files via upload
The data that support the findings of the paper, in particular the results from Figure 2. The repository contains 4 files: training_data_200_65_20_5, training_labels_200_20_5, test_data_20_65_20_5 and test_labels_20_20_5. The first file training_data_200_65_20_5 contains 65-dimensional points that are used for training, their labels are saved in the training_labels_200_20_5 file. Test data, that are also 65-dimensional are in the test_data_20_65_20_5 file, their labels are in test_labels_20_20_5. The matrices of the data and the labels are flattened - the original dimensions are saved in the name for further reconstruction. Each file contains 5 subsets of 20 different parameter setups (see data model definition in Section 4.1), for each setup 200 training points and 20 testing points are used.